Adding custom raw decoder, QueryRaw and StringBytes32 custom raw decoder
Mainly there is mixmatch of bytes32 types and strings. this introduces the capability of doing custom decoding.

+++ src/Nethereum.XUnitEthereumClients/EthereumClientIntegrationFixture.cs
@@ -108,6 +108,7 @@ namespace Nethereum.XUnitEthereumClients
 
         public void Dispose()
         {
+            if(!_process.HasExited)
             _process.Kill();
             Thread.Sleep(2000);
             DeleteData();
